Create testing-functions.py
unctions are one of the fundamental building blocks for writing well-factored, understandable code. Another benefit of functions is that we're able to package up code so that we don't need to think about the actual logic within the function. We can focus on the results, depending on our inputs. Viewing our functions as black boxes where we provide inputs and receive outputs means that we can also write automated tests to ensure that our functions (which might not even be written yet) behave the way that we expect. This is known as Test-Driven Development. We can use the assert statement to write some basic "tests" that utilize a function and ensure that the output is what we expect. We're going to work through the testing-functions.py file, writing and modifying functions to get the assertions throughout the file so that it doesn't throw errors.