Create using-generators.py
The range function returns an iterable object that makes it incredibly easy to iterate through a large list of integers without needing to create the list up front. In this hands-on lab, we're going to create the char_range generator that will do the same thing, except it will provide us with a range of characters based on Unicode code points for the starting character and ending character, with an optional step value.